face-id-platforma-master/backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
daphne==4.1.2
channels==4.1.0
channels-redis==4.2.0
celery==5.3.6
redis==5.0.1
psycopg2-binary>=2.9.10
django-cors-headers==4.3.1
django-redis==5.4.0
django-filter==24.1
django-storages==1.14.2
python-dotenv==1.0.1
Pillow>=11.0.0
gunicorn==21.2.0
numpy>=2.0.0
